third_party/blink/web_tests/external/wpt/editing/other/cloning-attributes-at-splitting-element.tentative-expected.txt

This is a testharness.js-based test.
Found 7 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Cloning attributes at splitting an element in contenteditable 1
  assert_equals: The new element created by splitting an element shouldn't have same id attribute value (Splitting <div id="splittee">) expected 1 but got 2
[FAIL] Cloning attributes at splitting an element in contenteditable 7
  assert_equals: The new element created by splitting an element shouldn't have same id attribute value (Splitting <li id="splittee">) expected 1 but got 2
[FAIL] Cloning attributes at splitting an element in contenteditable 13
  assert_equals: The new element created by splitting an element shouldn't have same id attribute value (Splitting <h3 id="p">) expected 1 but got 2
[FAIL] Cloning attributes at splitting an element in contenteditable 19
  assert_equals: The new element created by splitting an element shouldn't have same id attribute value (Splitting <dt id="splittee">) expected 1 but got 2
[FAIL] Cloning attributes at splitting an element in contenteditable 25
  assert_equals: The new element created by splitting an element shouldn't have same id attribute value (Splitting <dd id="splittee">) expected 1 but got 2
[FAIL] Cloning attributes at splitting an element in contenteditable 30
  assert_equals: The element instance returned by Document.getElementById shouldn't be changed after splitting the element (Splitting <div id="splittee-parent"> and <span id="splittee">) expected Element node <span>def</span> but got Element node <span id="splittee">abc</span>
[FAIL] Cloning attributes at splitting an element in contenteditable 33
  assert_equals: The new element created by splitting an element shouldn't have same id attribute value (splittee-parent) (Splitting <div id="splittee-parent"> and <span id="splittee">) expected 1 but got 2
Harness: the test ran to completion.
